Rename service to BLL in Contract
