+++ PracticalTest.Contracts/BLL/IServiceManager.cs
